[accumulators] Move code from feature/max_num_binning.hpp to .cpp
Result::covariance() and Result::accurate_covariance() seem to be
broken.
