+++ accumulators/src/feature/binning_analysis.cpp
@@ -7,6 +7,11 @@
 #include <alps/accumulators/feature/error.hpp>
 #include <alps/accumulators/feature/binning_analysis.hpp>
 
+#include <boost/preprocessor/tuple/to_seq.hpp>
+#include <boost/preprocessor/seq/for_each.hpp>
+
+#define ALPS_ACCUMULATOR_VALUE_TYPES_SEQ BOOST_PP_TUPLE_TO_SEQ(ALPS_ACCUMULATOR_VALUE_TYPES_SIZE, (ALPS_ACCUMULATOR_VALUE_TYPES))
+
 // DEBUG: to force boost assertion to be an exception, to work nicely with google test
 #define BOOST_ENABLE_ASSERT_HANDLER
 #include "boost/assert.hpp"
